moved tests from packages/test to owning packages
