+++ packages/custom-element/.npmignore
@@ -9,6 +9,7 @@ lerna-debug.log*
 
 node_modules
 example
+test
 dist-ssr
 *.local
 *.ts
