+++ packages/custom-element/package.json
@@ -12,8 +12,12 @@
     "@diax-js/common": "^0.0.2",
     "@diax-js/context": "^0.0.2"
   },
+  "devDependencies": {
+    "@diax-js/test": "../test"
+  },
   "scripts": {
-    "build": "tsc --build --verbose"
+    "build": "tsc --build --verbose",
+    "test": "vitest"
   },
   "keywords": [
     "web",
