+++ packages/custom-element/test/index.html
@@ -1,0 +1,10 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+</head>
+
+<body>
+</body>
+
+</html>
