+++ packages/form-element/package.json
@@ -13,8 +13,12 @@
     "@diax-js/context": "^0.0.2",
     "@diax-js/custom-element": "^0.0.2"
   },
+  "devDependencies": {
+    "@diax-js/test": "../test"
+  },
   "scripts": {
-    "build": "tsc --build --verbose"
+    "build": "tsc --build --verbose",
+    "test": "vitest"
   },
   "keywords": [
     "web",
