+++ packages/rendering-element/tsconfig.json
@@ -2,6 +2,7 @@
     "extends": "../../tsconfig-package.json",
     "include": [
         "src",
+        "test",
         "./index.ts",
         "./package.json"
     ],
