+++ packages/test/package.json
@@ -7,16 +7,8 @@
   "files": [
     "dist"
   ],
-  "dependencies": {
-    "@diax-js/common": "^0.0.2",
-    "@diax-js/context": "^0.0.2",
-    "@diax-js/custom-element": "^0.0.2",
-    "@diax-js/form-element": "^0.0.2",
-    "@diax-js/rendering-element": "^0.0.2"
-  },
   "scripts": {
-    "test": "vitest",
-    "build": "tsc --build"
+    "build": "tsc --build --verbose"
   },
   "keywords": [],
   "author": "Marek Miałkowski",
