+++ packages/test/src/utils.ts
@@ -1,90 +1,5 @@
-import {
-  CONTEXT,
-  ContextElement,
-  FormElementCallbacks,
-  HTMLElementCallbacks,
-  RenderingElementCallbacks,
-} from '@diax-js/common';
+import { CONTEXT, ContextElement } from '@diax-js/common';
 import { ElementContext } from '@diax-js/context';
-import { RenderingElement } from '@diax-js/rendering-element';
-import { FormElement } from '@diax-js/form-element';
-import { Element } from '@diax-js/custom-element';
-
-@Element('test-element')
-export class TestElement implements HTMLElementCallbacks {
-  static observedAttributes = ['test'];
-
-  spy = jest.fn();
-
-  connectedCallback(): void {
-    this.spy();
-  }
-  disconnectedCallback(): void {
-    this.spy();
-  }
-  attributeChangedCallback(name: string, oldValue: unknown, newValue: unknown): void {
-    this.spy(name, oldValue, newValue);
-  }
-  adoptedCallback(): void {
-    this.spy();
-  }
-}
-
-@FormElement('test-form-element')
-export class TestFormElement implements HTMLElementCallbacks, FormElementCallbacks {
-  static observedAttributes = ['test'];
-
-  spy = jest.fn();
-
-  connectedCallback(): void {
-    this.spy();
-  }
-  disconnectedCallback(): void {
-    this.spy();
-  }
-  attributeChangedCallback(name: string, oldValue: unknown, newValue: unknown): void {
-    this.spy(name, oldValue, newValue);
-  }
-  adoptedCallback(): void {
-    this.spy();
-  }
-  formAssociatedCallback(form: HTMLFormElement): void {
-    this.spy(form);
-  }
-  formDisabledCallback(disabled: boolean): void {
-    this.spy(disabled);
-  }
-  formResetCallback(): void {
-    this.spy();
-  }
-  formStateRestoreCallback(state: unknown, reason: 'autocomplete' | 'restore'): void {
-    this.spy(state, reason);
-  }
-}
-
-@RenderingElement('test-rendering-element')
-export class TestRenderingElement implements RenderingElementCallbacks {
-  static observedAttributes = ['test'];
-
-  spy = jest.fn();
-
-  connectedCallback(): void {
-    this.spy();
-  }
-  disconnectedCallback(): void {
-    this.spy();
-  }
-  attributeChangedCallback(name: string, oldValue: unknown, newValue: unknown): void {
-    this.spy(name, oldValue, newValue);
-  }
-  adoptedCallback(): void {
-    this.spy();
-  }
-
-  render(): void {
-    this.spy();
-  }
-}
 
 export function createContextElement(tagName: keyof HTMLElementTagNameMap): ContextElement {
   const element = document.createElement(tagName);
