+++ tsconfig-package.json
@@ -14,6 +14,10 @@
         "noImplicitOverride": true,
         "sourceMap": true,
         "declaration": true,
-        "declarationMap": true
+        "declarationMap": true,
+        "types": [
+            "vitest/globals",
+            "node"
+        ]
     },
 }
